refractor blog article express route
